requirements.txt
autopep8==1.5.7
coverage==5.5
cypress-django==0.1.0
dj-database-url==0.5.0
Django==2.1.4
django-dotenv==1.4.2
djangorestframework==3.9.0
flake8==3.9.2
gunicorn==19.9.0
mccabe==0.6.1
pep8==1.7.1
psycopg2-binary==2.9.1
pycodestyle==2.7.0
pyflakes==2.3.1
pytz==2018.7
toml==0.10.2
whitenoise==4.0
